Refresh browsers to force handshake
A race condition on real-time updates
(https://github.com/amark/gun/issues/259) was breaking the tests on gun
0.5.
To force a data handshake, the browsers are refreshed after the server
is restarted. Because of deadlines, support for automated tools like
phantomjs or any of the webdrivers has been broken, it will need to be
run manually:
- Start the test in the terminal
- Open a browser to the server's ip:port/ address
- Refresh the browsers every time you want to start the tests again
I know, lamesauce. Automation is way cooler.